refactor frontend sale modals
